fix(reporting): correct CurQTY type from decimal to int
- Change CurQTY property type in OverallTransferByProductionOrderRecord DTO to int
- Update data reader conversion from decimal to int for CurQTY in ReportingRepository
- Adjust Razor view to format CurQTY as integer instead of decimal with two decimals
- Ensure consistency of CurQTY across data layer and UI presentation

+++ Core/DTOs/OverallTransferByProductionOrderRecord.cs
@@ -9,7 +9,7 @@ namespace Web.Core.DTOs
         public string MaterialDescription { get; set; } = string.Empty;
         public string Batch { get; set; } = string.Empty;
         public int OrderQty { get; set; }
-        public decimal CurQTY { get; set; }
+        public int CurQTY { get; set; }
         public int IssueCount { get; set; }
         public int ReceiptCount { get; set; }
         public int Deviation { get; set; }
